assignments/docker/pyspark/project1/requirements.txt
boto3
pytest
python-dotenv
pyspark==3.5.1
